moved kubernetes package from internal to pkg
