+++ Dockerfile.backend
@@ -5,6 +5,6 @@ RUN CGO_ENABLED=0 GOOS=linux go build -a -installsuffix cgo -o kubeyaml ./cmd/se
 
 FROM alpine:3.8
 
-RUN mkdir -p /internal/kubernetes/data
+RUN mkdir -p /pkg/kubernetes/data
 COPY --from=0 /kubeyaml/kubeyaml /kubeyaml
 ENTRYPOINT [ "/kubeyaml" ]
