+++ pkg/kubernetes/data/swagger-1.10.json
