+++ pkg/kubernetes/data/swagger-1.11.json
