+++ pkg/kubernetes/data/swagger-1.12.json
