+++ pkg/kubernetes/data/swagger-1.8.json
