+++ pkg/kubernetes/data/swagger-1.9.json
