+++ pkg/kubernetes/data/swagger_OneEight.go
