+++ pkg/kubernetes/data/swagger_OneTen.go
