+++ pkg/kubernetes/data/swagger_OneTwelve.go
